+++ src/main/java/org/mmarini/routes/swing/v2/EdgePane.java
@@ -1,3 +1,29 @@
+//
+// Copyright (c) 2019 Marco Marini, marco.marini@mmarini.org
+//
+// Permission is hereby granted, free of charge, to any person
+// obtaining a copy of this software and associated documentation
+// files (the "Software"), to deal in the Software without
+// restriction, including without limitation the rights to use,
+// copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the
+// Software is furnished to do so, subject to the following
+// conditions:
+//
+// The above copyright notice and this permission notice shall be
+// included in all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+// EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
+// OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+// NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
+// HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
+// WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+// FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
+// OTHER DEALINGS IN THE SOFTWARE.
+//
+//   END OF TERMS AND CONDITIONS
+
 package org.mmarini.routes.swing.v2;
 
 import static org.mmarini.routes.swing.v2.SwingUtils.createJButton;
@@ -7,6 +33,8 @@ import java.awt.Component;
 import java.awt.GridBagConstraints;
 import java.awt.GridBagLayout;
 import java.awt.Insets;
+import java.awt.event.ActionEvent;
+import java.beans.PropertyChangeEvent;
 import java.text.NumberFormat;
 
 import javax.swing.BorderFactory;
@@ -19,6 +47,11 @@ import javax.swing.JTextField;
 import javax.swing.JToolBar;
 import javax.swing.SwingConstants;
 import javax.swing.text.NumberFormatter;
+
+import org.mmarini.routes.model.v2.MapEdge;
+
+import hu.akarnokd.rxjava3.swing.SwingObservable;
+import io.reactivex.rxjava3.core.Observable;
 
 /**
  *
@@ -35,21 +68,33 @@ public class EdgePane extends JPanel {
 	private final JTextField endField;
 	private final JButton browseBeginNodeButton;
 	private final JButton browseEndNodeButton;
+	private final JButton deleteButton;
+	private final Observable<ActionEvent> browseBeginObs;
+	private final Observable<ActionEvent> browseEndObs;
+	private final Observable<ActionEvent> deleteObs;
+	private final Observable<PropertyChangeEvent> priorityObs;
+	private final Observable<PropertyChangeEvent> speedLimitObs;
 
 	/**
-	     *
-	     */
+	 *
+	 */
 	public EdgePane() {
 		priorityField = new JFormattedTextField(new NumberFormatter(NumberFormat.getIntegerInstance()));
 		speedLimitField = new JFormattedTextField(new NumberFormatter(NumberFormat.getNumberInstance()));
 		incomeQueueField = new JFormattedTextField(new NumberFormatter(NumberFormat.getIntegerInstance()));
 		distanceField = new JFormattedTextField(new NumberFormatter(NumberFormat.getNumberInstance()));
-		nameField = new JTextField(20);
-		beginField = new JTextField(10);
-		endField = new JTextField(10);
+		nameField = new JTextField(6);
+		beginField = new JTextField(6);
+		endField = new JTextField(6);
 		browseBeginNodeButton = createJButton("EdgePane.browseBeginNodeAction"); //$NON-NLS-1$
 		browseEndNodeButton = createJButton("EdgePane.browseEndNodeAction"); //$NON-NLS-1$
+		deleteButton = createJButton("EdgePane.deleteAction"); //$NON-NLS-1$
 
+		browseBeginObs = SwingObservable.actions(browseBeginNodeButton);
+		browseEndObs = SwingObservable.actions(browseBeginNodeButton);
+		deleteObs = SwingObservable.actions(deleteButton);
+		priorityObs = SwingObservable.propertyChange(priorityField, "value");
+		speedLimitObs = SwingObservable.propertyChange(speedLimitField, "value");
 		init().createContent();
 	}
 
@@ -282,6 +327,41 @@ public class EdgePane extends JPanel {
 	}
 
 	/**
+	 * @return the browseBeginObs
+	 */
+	public Observable<ActionEvent> getBrowseBeginObs() {
+		return browseBeginObs;
+	}
+
+	/**
+	 * @return the browseEndObs
+	 */
+	public Observable<ActionEvent> getBrowseEndObs() {
+		return browseEndObs;
+	}
+
+	/**
+	 * @return the deleteObs
+	 */
+	public Observable<ActionEvent> getDeleteObs() {
+		return deleteObs;
+	}
+
+	/**
+	 * @return the priorityObs
+	 */
+	public Observable<PropertyChangeEvent> getPriorityObs() {
+		return priorityObs;
+	}
+
+	/**
+	 * @return the speedLimitObs
+	 */
+	public Observable<PropertyChangeEvent> getSpeedLimitObs() {
+		return speedLimitObs;
+	}
+
+	/**
 	 * @return
 	 *
 	 */
@@ -321,6 +401,20 @@ public class EdgePane extends JPanel {
 		beginField.setEditable(false);
 		endField.setEditable(false);
 
+		return this;
+	}
+
+	/**
+	 *
+	 * @param edge
+	 */
+	public EdgePane setEdge(final MapEdge edge) {
+		priorityField.setValue(edge.getPriority());
+		speedLimitField.setValue(edge.getSpeedLimit() * 3.6);
+		beginField.setText(edge.getBegin().getShortName());
+		endField.setText(edge.getEnd().getShortName());
+		distanceField.setValue(edge.getDistance());
+		nameField.setText(edge.getShortName());
 		return this;
 	}
 }