+++ src/main/java/org/mmarini/routes/swing/v2/MapNodePane.java
@@ -33,6 +33,8 @@ import java.awt.Component;
 import java.awt.GridBagConstraints;
 import java.awt.GridBagLayout;
 import java.awt.Insets;
+import java.awt.event.ActionEvent;
+import java.awt.geom.Point2D;
 
 import javax.swing.BorderFactory;
 import javax.swing.Box;
@@ -46,6 +48,11 @@ import javax.swing.JToolBar;
 import javax.swing.SwingConstants;
 import javax.swing.text.NumberFormatter;
 
+import org.mmarini.routes.model.v2.MapNode;
+
+import hu.akarnokd.rxjava3.swing.SwingObservable;
+import io.reactivex.rxjava3.core.Observable;
+
 /**
  *
  */
@@ -55,21 +62,23 @@ public class MapNodePane extends JPanel {
 	private final JTextField nameField;
 	private final JFormattedTextField xField;
 	private final JFormattedTextField yField;
-
-	private final JLabel nameLabel;
-	private final JLabel xLabel;
-	private final JLabel yLabel;
+	private final JButton changeButton;
+	private final JButton deleteButton;
+	private final Observable<ActionEvent> changeObs;
+	private final Observable<ActionEvent> deleteObs;
 
 	/**
 	 *
 	 */
 	public MapNodePane() {
-		nameField = new JTextField(10);
+		nameField = new JTextField(6);
 		xField = new JFormattedTextField(new NumberFormatter());
 		yField = new JFormattedTextField(new NumberFormatter());
-		nameLabel = new JLabel(Messages.getString("MapNodePane.name.label")); //$NON-NLS-1$
-		xLabel = new JLabel(Messages.getString("MapNodePane.x.label")); //$NON-NLS-1$
-		yLabel = new JLabel(Messages.getString("MapNodePane.y.label")); //$NON-NLS-1$
+		changeButton = createJButton("MapNodePane.changeAction"); //$NON-NLS-1$
+		deleteButton = createJButton("MapNodePane.deleteAction"); //$NON-NLS-1$
+
+		changeObs = SwingObservable.actions(changeButton);
+		deleteObs = SwingObservable.actions(changeButton);
 
 		setBorder(BorderFactory.createTitledBorder(Messages.getString("MapNodePane.title"))); //$NON-NLS-1$
 
@@ -105,7 +114,6 @@ public class MapNodePane extends JPanel {
 		cons.insets = new Insets(2, 2, 2, 2);
 
 		Component c;
-		c = nameLabel;
 		cons.gridx = 0;
 		cons.gridy = 0;
 		cons.gridwidth = 1;
@@ -114,10 +122,10 @@ public class MapNodePane extends JPanel {
 		cons.fill = GridBagConstraints.NONE;
 		cons.weightx = 0;
 		cons.weighty = 0;
+		c = new JLabel(Messages.getString("MapNodePane.name.label")); //$NON-NLS-1$
 		layout.setConstraints(c, cons);
 		pane.add(c);
 
-		c = nameField;
 		cons.gridx = 1;
 		cons.gridy = 0;
 		cons.gridwidth = 1;
@@ -126,10 +134,10 @@ public class MapNodePane extends JPanel {
 		cons.fill = GridBagConstraints.HORIZONTAL;
 		cons.weightx = 0;
 		cons.weighty = 0;
+		c = nameField;
 		layout.setConstraints(c, cons);
 		pane.add(c);
 
-		c = Box.createGlue();
 		cons.gridx = 2;
 		cons.gridy = 0;
 		cons.gridwidth = 1;
@@ -138,58 +146,58 @@ public class MapNodePane extends JPanel {
 		cons.fill = GridBagConstraints.NONE;
 		cons.weightx = 1;
 		cons.weighty = 0;
-		layout.setConstraints(c, cons);
-		pane.add(c);
-
-		c = xLabel;
-		cons.gridx = 0;
-		cons.gridy = 1;
-		cons.gridwidth = 1;
-		cons.gridheight = 1;
-		cons.anchor = GridBagConstraints.EAST;
-		cons.fill = GridBagConstraints.NONE;
-		cons.weightx = 0;
-		cons.weighty = 0;
-		layout.setConstraints(c, cons);
-		pane.add(c);
-
-		c = xField;
-		cons.gridx = 1;
-		cons.gridy = 1;
-		cons.gridwidth = 1;
-		cons.gridheight = 1;
-		cons.anchor = GridBagConstraints.WEST;
-		cons.fill = GridBagConstraints.NONE;
-		cons.weightx = 0;
-		cons.weighty = 0;
-		layout.setConstraints(c, cons);
-		pane.add(c);
-
-		c = yLabel;
-		cons.gridx = 0;
-		cons.gridy = 2;
-		cons.gridwidth = 1;
-		cons.gridheight = 1;
-		cons.anchor = GridBagConstraints.EAST;
-		cons.fill = GridBagConstraints.NONE;
-		cons.weightx = 0;
-		cons.weighty = 0;
-		layout.setConstraints(c, cons);
-		pane.add(c);
-
-		c = yField;
-		cons.gridx = 1;
-		cons.gridy = 2;
-		cons.gridwidth = 1;
-		cons.gridheight = 1;
-		cons.anchor = GridBagConstraints.WEST;
-		cons.fill = GridBagConstraints.NONE;
-		cons.weightx = 0;
-		cons.weighty = 0;
-		layout.setConstraints(c, cons);
-		pane.add(c);
-
 		c = Box.createGlue();
+		layout.setConstraints(c, cons);
+		pane.add(c);
+
+		cons.gridx = 0;
+		cons.gridy = 1;
+		cons.gridwidth = 1;
+		cons.gridheight = 1;
+		cons.anchor = GridBagConstraints.EAST;
+		cons.fill = GridBagConstraints.NONE;
+		cons.weightx = 0;
+		cons.weighty = 0;
+		c = new JLabel(Messages.getString("MapNodePane.x.label")); //$NON-NLS-1$
+		layout.setConstraints(c, cons);
+		pane.add(c);
+
+		cons.gridx = 1;
+		cons.gridy = 1;
+		cons.gridwidth = 1;
+		cons.gridheight = 1;
+		cons.anchor = GridBagConstraints.WEST;
+		cons.fill = GridBagConstraints.NONE;
+		cons.weightx = 0;
+		cons.weighty = 0;
+		c = xField;
+		layout.setConstraints(c, cons);
+		pane.add(c);
+
+		cons.gridx = 0;
+		cons.gridy = 2;
+		cons.gridwidth = 1;
+		cons.gridheight = 1;
+		cons.anchor = GridBagConstraints.EAST;
+		cons.fill = GridBagConstraints.NONE;
+		cons.weightx = 0;
+		cons.weighty = 0;
+		c = new JLabel(Messages.getString("MapNodePane.y.label")); //$NON-NLS-1$
+		layout.setConstraints(c, cons);
+		pane.add(c);
+
+		cons.gridx = 1;
+		cons.gridy = 2;
+		cons.gridwidth = 1;
+		cons.gridheight = 1;
+		cons.anchor = GridBagConstraints.WEST;
+		cons.fill = GridBagConstraints.NONE;
+		cons.weightx = 0;
+		cons.weighty = 0;
+		c = yField;
+		layout.setConstraints(c, cons);
+		pane.add(c);
+
 		cons.gridx = 0;
 		cons.gridy = 3;
 		cons.gridwidth = 2;
@@ -198,6 +206,7 @@ public class MapNodePane extends JPanel {
 		cons.fill = GridBagConstraints.NONE;
 		cons.weightx = 0;
 		cons.weighty = 1;
+		c = Box.createGlue();
 		layout.setConstraints(c, cons);
 		pane.add(c);
 
@@ -209,11 +218,35 @@ public class MapNodePane extends JPanel {
 	 */
 	private JComponent createToolBar() {
 		final JToolBar toolbar = new JToolBar();
-		final JButton changeButton = createJButton("MapNodePane.changeAction"); //$NON-NLS-1$
-		final JButton deleteButton = createJButton("MapNodePane.deleteAction"); //$NON-NLS-1$
 
 		toolbar.add(changeButton);
 		toolbar.add(deleteButton);
 		return toolbar;
+	}
+
+	/**
+	 * @return the changeObs
+	 */
+	public Observable<ActionEvent> getChangeObs() {
+		return changeObs;
+	}
+
+	/**
+	 * @return the deleteObs
+	 */
+	public Observable<ActionEvent> getDeleteObs() {
+		return deleteObs;
+	}
+
+	/**
+	 *
+	 * @param node
+	 */
+	public MapNodePane setNode(final MapNode node) {
+		final Point2D location = node.getLocation();
+		xField.setValue(location.getX());
+		yField.setValue(location.getY());
+		nameField.setText(node.getShortName());
+		return this;
 	}
 }