+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Issue #28: Zoom in / out
 - Issue #29: Head up display
 - Issue #30: Explorer panel
+- Issue #31: Detail info panel
 - Issue #32: Scroll panel
 - Issue #35: Center map to cursor